try to handle the INT_MIN as input for _printf

+++ main.h
@@ -8,8 +8,8 @@
 int spformat(char);
 int print_reverse(int);
 size_t _pow(size_t, int);
-int print_int(size_t n, char);
-int print_bin(int b);
+int print_int(long int n, char);
+int print_bin(long int b);
 int _printf(const char *format, ...);
 int print_arg(va_list *args, char c);
 #endif
